requirements.txt
python-telegram-bot[job-queue]==21.5
pytz
httpx
